# frontend/apps/docs/src/pages/api-reference/assets/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Footprint API"
    },
    "servers": [
        {
            "url": "//api.onefootprint.com"
        }
    ],
    "paths": {
...
        "/vault_proxy/jit": {
            "post": {
                "tags": [
                    "VaultProxy",
                    "PublicApi"
                ],
                "description": "Invoke the vault proxy 'just-in-time' (JIT) to securely send and receive data to a target destination. Please contact support to enable this API.",
                "parameters": [
                    {
                        "in": "header",
                        "name": "x-fp-proxy-target-url",
                        "description": "Target proxy destination URL.",
                        "required": true,
                        "schema": {
                            "example": "https://acmebankprocessor.com/api/process_transaction",
                            "type": "string"
                        },
                        "example": "https://acmebankprocessor.com/api/process_transaction",
                        "style": "simple"
                    },
                    {
                        "in": "header",
                        "name": "x-fp-proxy-method",
                        "description": "HTTP Method VERB for the proxy destination request (defaults to POST).",
                        "schema": {
                            "type": "string"
                        },
                        "style": "simple"
                    },
                    {
                        "in": "header",
                        "name": "x-fp-path-and-query",
                        "description": "Egress destination URL path and query string to append.",
                        "schema": {
                            "type": "string"
                        },
                        "style": "simple"
                    },
                    {
                        "in": "header",
                        "name": "x-fp-proxy-ingress-content-type",
                        "description": "Content-type for the proxy ingress.",
                        "schema": {
                            "type": "string"
                        },
                        "style": "simple"
                    },
                    {
                        "in": "header",
                        "name": "x-fp-proxy-access-reason",
                        "description": "Access reason for any egress decryption operations during the proxy request.",
                        "schema": {
                            "type": "string"
                        },
                        "style": "simple"
                    },
                    {
                        "in": "header",
                        "name": "x-fp-id",
                        "description": "When proxy requests are on behalf of a single footprint vault, you can omit the `fp_id_` prefix on token identifiers, and just use `id.x` or `custom.y` instead of `fp_id_xyz.id.x` or `fp_id_xyz.custom.y`.  Similarly, if specifying proxy configuration ingress rules from a stored configuration the corresponding token must be assigned just-in-time via a headers.  i.e: `x-fp-id: fp_id_abc`",
                        "schema": {
                            "type": "string"
                        },
                        "style": "simple"
                    },
                    {
                        "in": "header",
                        "name": "x-fp-proxy-ingress-rule",
                        "description": "Configure one more ingress rules.  x-fp-proxy-ingress-rule: fp_id_abc.custom.credit_card_number=$.data.card.number x-fp-proxy-ingress-rule: fp_id_abc.custom.credit_card_exp=$.data.card.expiration x-fp-proxy-ingress-rule: fp_id_abc.custom.credit_card_cvc=$.data.card.security_code",
                        "schema": {
                            "type": "string"
                        },
                        "style": "simple"
                    },
                    {
                        "in": "header",
                        "name": "x-fp-proxy-fwd-",
                        "description": "Headers with this `x-fp-proxy-fwd-*` prefix are forwarded to the proxy egress, with the prefix stripped. For example `x-fp-proxy-fwd-MYHEADER: hello world` sends `MYHEADER: hello world` to the destination.",
                        "schema": {
                            "type": "string"
                        },
                        "style": "simple"
                    },
                    {
                        "in": "header",
                        "name": "x-fp-proxy-client-cert",
                        "description": "Base64 encoded PEM client certificate to use (required if using key).",
                        "schema": {
                            "type": "string"
                        },
                        "style": "simple"
                    },
                    {
                        "in": "header",
                        "name": "x-fp-proxy-client-key",
                        "description": "Base64 encoded PEM client key to use (required if using cert).",
                        "schema": {
                            "type": "string"
                        },
                        "style": "simple"
                    },
                    {
                        "in": "header",
                        "name": "x-fp-proxy-pin-cert",
                        "description": "Configure one or more base64 encoded PEM server certificates to validate and pin proxy destination TLS connections.",
                        "schema": {
                            "type": "string"
                        },
                        "style": "simple"
                    }
                ],
                "requestBody": {
                    "description": "Raw body",
                    "content": {
                        "application/json": {
                            "schema": {
                                "example": "raw data",
                                "description": "Raw body",
                                "type": "string"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {},
                "security": [
                    {
                        "Secret API Key": [
                            "preview:vault_proxy_jit"
                        ]
                    }
                ]
            }
        },
...
        "/vault_proxy/{id}": {
            "post": {
                "tags": [
                    "VaultProxy",
                    "PublicApi"
                ],
                "description": "Invoke the vault proxy by configuration id to securely send and receive data to a target destination. Please contact support to enable this API.",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "style": "simple"
                    },
                    {
                        "in": "header",
                        "name": "x-fp-proxy-method",
                        "description": "HTTP Method VERB for the proxy destination request (defaults to POST).",
                        "schema": {
                            "type": "string"
                        },
                        "style": "simple"
                    },
                    {
                        "in": "header",
                        "name": "x-fp-path-and-query",
                        "description": "Egress destination URL path and query string to append.",
                        "schema": {
                            "type": "string"
                        },
                        "style": "simple"
                    },
                    {
                        "in": "header",
                        "name": "x-fp-proxy-ingress-content-type",
                        "description": "Content-type for the proxy ingress.",
                        "schema": {
                            "type": "string"
                        },
                        "style": "simple"
                    },
                    {
                        "in": "header",
                        "name": "x-fp-proxy-access-reason",
                        "description": "Access reason for any egress decryption operations during the proxy request.",
                        "schema": {
                            "type": "string"
                        },
                        "style": "simple"
                    },
                    {
                        "in": "header",
                        "name": "x-fp-id",
                        "description": "When proxy requests are on behalf of a single footprint vault, you can omit the `fp_id_` prefix on token identifiers, and just use `id.x` or `custom.y` instead of `fp_id_xyz.id.x` or `fp_id_xyz.custom.y`.  Similarly, if specifying proxy configuration ingress rules from a stored configuration the corresponding token must be assigned just-in-time via a headers.  i.e: `x-fp-id: fp_id_abc`",
                        "schema": {
                            "type": "string"
                        },
                        "style": "simple"
                    },
                    {
                        "in": "header",
                        "name": "x-fp-proxy-ingress-rule",
                        "description": "Configure one more ingress rules.  x-fp-proxy-ingress-rule: fp_id_abc.custom.credit_card_number=$.data.card.number x-fp-proxy-ingress-rule: fp_id_abc.custom.credit_card_exp=$.data.card.expiration x-fp-proxy-ingress-rule: fp_id_abc.custom.credit_card_cvc=$.data.card.security_code",
                        "schema": {
                            "type": "string"
                        },
                        "style": "simple"
                    },
                    {
                        "in": "header",
                        "name": "x-fp-proxy-fwd-",
                        "description": "Headers with this `x-fp-proxy-fwd-*` prefix are forwarded to the proxy egress, with the prefix stripped. For example `x-fp-proxy-fwd-MYHEADER: hello world` sends `MYHEADER: hello world` to the destination.",
                        "schema": {
                            "type": "string"
                        },
                        "style": "simple"
                    },
                    {
                        "in": "header",
                        "name": "x-fp-proxy-client-cert",
                        "description": "Base64 encoded PEM client certificate to use (required if using key).",
                        "schema": {
                            "type": "string"
                        },
                        "style": "simple"
                    },
                    {
                        "in": "header",
                        "name": "x-fp-proxy-client-key",
                        "description": "Base64 encoded PEM client key to use (required if using cert).",
                        "schema": {
                            "type": "string"
                        },
                        "style": "simple"
                    },
                    {
                        "in": "header",
                        "name": "x-fp-proxy-pin-cert",
                        "description": "Configure one or more base64 encoded PEM server certificates to validate and pin proxy destination TLS connections.",
                        "schema": {
                            "type": "string"
                        },
                        "style": "simple"
                    }
                ],
                "requestBody": {
                    "description": "Raw body",
                    "content": {
                        "application/json": {
                            "schema": {
                                "example": "raw data",
                                "description": "Raw body",
                                "type": "string"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {},
                "security": [
                    {
                        "Secret API Key": [
                            "preview:vault_proxy"
                        ]
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "Client Token": {
                "type": "apiKey",
                "in": "header",
                "name": "X-Fp-Authorization",
                "description": "Short-lived client token to perform actions for a given user."
            },
            "Secret API Key": {
                "type": "apiKey",
                "in": "header",
                "name": "X-Footprint-Secret-Key",
                "description": "Secret API key. You can create and view your API keys in the dashboard. This key should never be sent to your client and should be treated as a secret on your server."
            }
        },
...
    },
...
}